Non-working wgpu impl

+++ src/cli.rs
@@ -11,22 +11,22 @@ pub struct Cli {
     pub output: PathBuf,
 
     #[arg(long = "qrh", help = "The quantum efficiency of the red channel at the hydrogen-alpha wavelength (656.3 nm)")]
-    pub red_ha_qe: f64,
+    pub red_ha_qe: f32,
 
     #[arg(long = "qgh", help = "The quantum efficiency of the green channel at the hydrogen-alpha wavelength (656.3 nm)")]
-    pub green_ha_qe: f64,
+    pub green_ha_qe: f32,
 
     #[arg(long = "qbh", help = "The quantum efficiency of the blue channel at the hydrogen-alpha wavelength (656.3 nm)")]
-    pub blue_ha_qe: f64,
+    pub blue_ha_qe: f32,
 
     #[arg(long = "qro", help = "The quantum efficiency of the red channel at the OIII wavelength (500.7 nm)")]
-    pub red_oiii_qe: f64,
+    pub red_oiii_qe: f32,
 
     #[arg(long = "qgo", help = "The quantum efficiency of the green channel at the OIII wavelength (500.7 nm)")]
-    pub green_oiii_qe: f64,
+    pub green_oiii_qe: f32,
 
     #[arg(long = "qbo", help = "The quantum efficiency of the blue channel at the OIII wavelength (500.7 nm)")]
-    pub blue_oiii_qe: f64,
+    pub blue_oiii_qe: f32,
 
     #[arg(short, long, default_value_t = 100, help = "Population size for the genetic algorithm")]
     pub population_size: usize,
@@ -38,8 +38,8 @@ pub struct Cli {
     pub elitism: usize,
 
     #[arg(short = 's', long, default_value_t = 0.5, help = "Initial standard deviation for mutation")]
-    pub initial_std: f64,
+    pub initial_std: f32,
     
     #[arg(short, long, default_value_t = 0.1, help = "Decay rate for mutation standard deviation")]
-    pub decay_rate: f64,
+    pub decay_rate: f32,
 }
